Create Read me
This project is part of the CPTS 122 final project.
I and 3 more member, Brandon Truong, Micah Moody, Aayush Aesthresha, worked together on this project.
A little detail or tutorial about this project, it is similar to other gaming projects (crossy road) just in our case we decided to take that inspiration and make it our own ways. We implemented SFML and inheritance in this project as part of the project requirement.